requirements.txt
fastapi>=0.128.0
google-generativeai>=0.8.6
starlette>=0.52.1,<0.53.0
uvicorn[standard]>=0.40.0
python-dotenv>=1.2.1
prometheus-client>=0.24.1
gunicorn>=24.1.1
protobuf<6
